test-engineering/contract/pyproject.toml
[tool.black]
line-length = 88


[tool.isort]
profile = "black"
skip_gitignore = true


[tool.mypy]
python_version = "3.11"
disallow_untyped_calls = true
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_error_codes = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
warn_unreachable = true


[tool.poetry]
name = "contract"
version = "0.1.0"
description = "Contract tests for the Contile project"
authors = ["Mozilla"]
license = "Mozilla Public License Version 2.0"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0.3"
pytest = "^7.4.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
types-pyyaml = "^6.0.12.10"
types-requests = "^2.31.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.4.1"

[tool.poetry.group.partner.dependencies]
fastapi = "^0.100.0"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.23.1"}


[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
